Add files via upload
In this current state of the game, I added in a "great food" (yellow) that will give you 3 points when it is collected. As of now, it will spawn right away and right after it is eaten. I am working on timing it so that the yellow food only spawns every 30 seconds or so. Also, with the bad (red) food, that is also stationary. I am working on making the red food move around ever 5-10 seconds. Updates will be coming soon!